Fix global extension schema `Enum` types not retaining their `::` prefix (athena-framework/athena#639)
## Context

If a bundle schema defined an enum configuration property like:
```cr
property failed_validation_status : ::HTTP::Status = :unprocessable_entity
```
the `::` prefix would be lost when it was converted into
`HTTP::Status.new(:unprocessable_entity)`, potentially causing it to not
resolve to the correct type. This PR captures and re-adds the `::`
prefix if is a global type.

## Changelog

- Fix global extension schema `Enum` types not retaining their `::`
prefix